requirements.txt
Faker==24.8.0
PySocks==1.7.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0